Add monochrome logo, change normal logo to svg/webp, and circle
Svg allows it to scale to any screen density/usage, and take up less space. Unfortunately my photopea refuses to export the monochrome version as an svg without filling in the text 'galaxy' which looks quite bad, so the monochrome logos are webps instead.

Legacy icon still has to be an image format for older versions of android. Interestingly android 4.0 (this app's minimum version) and up support webp, and only 5.0 svgs (although as logos idk? maybe only since adaptive icons) so webps are used instead of png for space saving. I've also changed the legacy icon to be a circle so only android 7.1 and below it's a circle instead of a square which I think looks better on the OS.